+++ src/main/java/core/basesyntax/strategy/impl/DefaultDiscountService.java
@@ -5,7 +5,7 @@ import core.basesyntax.strategy.DiscountService;
 public class DefaultDiscountService implements DiscountService {
 
     @Override
-    public double getDiscount(int price) {
-        return (double) price;
+    public double getDiscount(double price) {
+        return price;
     }
 }
