+++ src/main/java/core/basesyntax/strategy/impl/NewYearDiscountService.java
@@ -4,7 +4,7 @@ import core.basesyntax.strategy.DiscountService;
 
 public class NewYearDiscountService implements DiscountService {
     @Override
-    public double getDiscount(int price) {
+    public double getDiscount(double price) {
         return price * 0.8;
     }
 }
